work fixing up filter generation.
