+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/apirequest/binders/DefaultDimensionFieldGenerator.java
@@ -1,0 +1,144 @@
+package com.yahoo.bard.webservice.web.apirequest.binders;
+
+import static com.yahoo.bard.webservice.web.ErrorMessageFormat.DIMENSION_FIELDS_UNDEFINED;
+
+import com.yahoo.bard.webservice.data.dimension.Dimension;
+import com.yahoo.bard.webservice.data.dimension.DimensionDictionary;
+import com.yahoo.bard.webservice.data.dimension.DimensionField;
+import com.yahoo.bard.webservice.logging.RequestLog;
+import com.yahoo.bard.webservice.logging.TimedPhase;
+import com.yahoo.bard.webservice.table.LogicalTable;
+import com.yahoo.bard.webservice.util.Incubating;
+import com.yahoo.bard.webservice.util.StreamUtils;
+import com.yahoo.bard.webservice.web.BadApiRequestException;
+import com.yahoo.bard.webservice.web.DimensionFieldSpecifierKeywords;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.Arrays;
+import java.util.LinkedHashMap;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import javax.validation.constraints.NotNull;
+import javax.ws.rs.core.PathSegment;
+
+@Incubating
+public class DefaultDimensionFieldGenerator implements DimensionFieldGenerator {
+    private static final Logger LOG = LoggerFactory.getLogger(DefaultDimensionFieldGenerator.class);
+
+    @Override
+    public void validateDimensionFields(
+            List<PathSegment> apiDimensionPathSegments,
+            LinkedHashMap<Dimension, LinkedHashSet<DimensionField>> perDimensionFields,
+            LinkedHashSet<Dimension> dimensions,
+            LogicalTable logicalTable,
+            DimensionDictionary dimensionDictionary
+    ) {
+        // by default no validation required
+    }
+
+    /**
+     * Extracts the list of dimensions from the url dimension path segments and "show" matrix params and generates a map
+     * of dimension to dimension fields which needs to be annotated on the response.
+     * <p>
+     * If no "show" matrix param has been set, it returns the default dimension fields configured for the dimension.
+     *
+     * @param apiDimensionPathSegments  Path segments for the dimensions
+     * @param dimensionDictionary  Dimension dictionary to look the dimensions up in
+     *
+     * @return A map of dimension to requested dimension fields
+     */
+    @Override
+    public LinkedHashMap<Dimension, LinkedHashSet<DimensionField>> generateDimensionFields(
+            @NotNull List<PathSegment> apiDimensionPathSegments,
+            @NotNull DimensionDictionary dimensionDictionary
+    ) {
+        try (TimedPhase timer = RequestLog.startTiming("GeneratingDimensionFields")) {
+            return apiDimensionPathSegments.stream()
+                    .filter(pathSegment -> !pathSegment.getPath().isEmpty())
+                    .collect(Collectors.toMap(
+                            pathSegment -> dimensionDictionary.findByApiName(pathSegment.getPath()),
+                            pathSegment -> bindShowClause(pathSegment, dimensionDictionary),
+                            StreamUtils::orderedSetMerge,
+                            LinkedHashMap::new
+                    ));
+        }
+    }
+
+    /**
+     * Given a path segment, bind the fields specified in it's "show" matrix parameter for the dimension specified in
+     * the path segment's path.
+     *
+     * @param pathSegment  Path segment to bind from
+     * @param dimensionDictionary  Dimension dictionary to look the dimension up in
+     *
+     * @return the set of bound DimensionFields specified in the show clause
+     * @throws BadApiRequestException if any of the specified fields are not valid for the dimension
+     */
+    private LinkedHashSet<DimensionField> bindShowClause(
+            PathSegment pathSegment,
+            DimensionDictionary dimensionDictionary
+    )
+            throws BadApiRequestException {
+        Dimension dimension = dimensionDictionary.findByApiName(pathSegment.getPath());
+        List<String> showFields = pathSegment.getMatrixParameters().entrySet().stream()
+                .filter(entry -> entry.getKey().equals("show"))
+                .flatMap(entry -> entry.getValue().stream())
+                .flatMap(s -> Arrays.stream(s.split(",")))
+                .collect(Collectors.toList());
+
+        if (showFields.size() == 1 && showFields.contains(DimensionFieldSpecifierKeywords.ALL.toString())) {
+            // Show all fields
+            return dimension.getDimensionFields();
+        } else if (showFields.size() == 1 && showFields.contains(DimensionFieldSpecifierKeywords.NONE.toString())) {
+            // Show no fields
+            return new LinkedHashSet<>();
+        } else if (!showFields.isEmpty()) {
+            // Show the requested fields
+            return bindDimensionFields(dimension, showFields);
+        } else {
+            // Show the default fields
+            return dimension.getDefaultDimensionFields();
+        }
+    }
+
+    /**
+     * Given a Dimension and a set of DimensionField names, bind the names to the available dimension fields of the
+     * dimension.
+     *
+     * @param dimension  Dimension to bind the fields for
+     * @param showFields  Names of the fields to bind
+     *
+     * @return the set of DimensionFields for the names
+     * @throws BadApiRequestException if any of the names are not dimension fields on the dimension
+     */
+    private LinkedHashSet<DimensionField> bindDimensionFields(Dimension dimension, List<String> showFields)
+            throws BadApiRequestException {
+        Map<String, DimensionField> dimensionNameToFieldMap = dimension.getDimensionFields().stream()
+                .collect(StreamUtils.toLinkedDictionary(DimensionField::getName));
+
+        LinkedHashSet<DimensionField> dimensionFields = new LinkedHashSet<>();
+        Set<String> invalidDimensionFields = new LinkedHashSet<>();
+        for (String field : showFields) {
+            if (dimensionNameToFieldMap.containsKey(field)) {
+                dimensionFields.add(dimensionNameToFieldMap.get(field));
+            } else {
+                invalidDimensionFields.add(field);
+            }
+        }
+
+        if (!invalidDimensionFields.isEmpty()) {
+            LOG.debug(DIMENSION_FIELDS_UNDEFINED.logFormat(invalidDimensionFields, dimension.getApiName()));
+            throw new BadApiRequestException(DIMENSION_FIELDS_UNDEFINED.format(
+                    invalidDimensionFields,
+                    dimension.getApiName()
+            ));
+        }
+        return dimensionFields;
+    }
+}
