+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/apirequest/binders/FilterFactory.java
@@ -7,6 +7,7 @@ import com.yahoo.bard.webservice.data.dimension.DimensionField;
 import com.yahoo.bard.webservice.util.Incubating;
 import com.yahoo.bard.webservice.web.ApiFilter;
 import com.yahoo.bard.webservice.web.FilterOperation;
+import com.yahoo.bard.webservice.web.apirequest.binders.FilterGenerationUtils.FilterComponents;
 
 import java.util.ArrayList;
 import java.util.LinkedHashSet;
@@ -18,6 +19,8 @@ import java.util.function.Predicate;
 /**
  * Filter factory provides a chain of responsibility for building ApiFilters.
  */
+//TODO This class feels overly complicated. Do we really need a chain of responsibility for building ApiFilters? I
+// think an interface for producing an ApiFilter with a default implementation is sufficient.
 @Incubating
 public class FilterFactory {
 
@@ -51,37 +54,6 @@ public class FilterFactory {
         );
     }
 
-
-    /**
-     * Data object for collecting the bound arguments needed to build an ApiFilter.
-     */
-    public static class FilterComponents {
-        public final Dimension dimension;
-        public final DimensionField dimensionField;
-        public final FilterOperation operation;
-        public final List<String> values;
-
-        /**
-         * Constructor.
-         *
-         * @param dimension  Dimension for ApiFilter.
-         * @param field  Field for ApiFilter.
-         * @param operation  Operation for ApiFilter.
-         * @param values  Values for ApiFilters.
-         */
-        public FilterComponents(
-                Dimension dimension,
-                DimensionField field,
-                FilterOperation operation,
-                List<String> values
-        ) {
-            this.dimension = dimension;
-            this.dimensionField = field;
-            this.operation = operation;
-            this.values = values;
-        }
-    }
-
     /**
      * Constructor.
      *
@@ -89,7 +61,7 @@ public class FilterFactory {
      * @param defaultFilterFactory  Default factory method for building ApiFilters.
      */
     protected FilterFactory(
-            List<Map.Entry<Predicate<FilterComponents>, FilterFactoryFunction>> filterFactoryProviders,
+            List<Map.Entry<Predicate<FilterGenerationUtils.FilterComponents>, FilterFactoryFunction>> filterFactoryProviders,
             FilterFactoryFunction defaultFilterFactory
     ) {
         this.filterFactoryProviders = filterFactoryProviders;
