+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/filterbuilders/ConjunctionDruidFilterBuilderSpec.groovy
@@ -14,7 +14,6 @@ import com.yahoo.bard.webservice.druid.model.filter.Filter
 import com.yahoo.bard.webservice.druid.model.filter.SelectorFilter
 import com.yahoo.bard.webservice.web.ApiFilter
 import com.yahoo.bard.webservice.web.apirequest.binders.FilterBinders
-import com.yahoo.bard.webservice.web.filters.ApiFilters
 
 import org.apache.commons.lang3.tuple.Pair
